refactor code to explit keycloakd and did servers
